Some adds
These three have different functions. The function add-on now allows a
social menu to be linked for social media inclusion, the Enqueue allows
for google fonts to be properly added into css, and font awesome will
be for certain images.
Header- took out link for slider, will add in another, created link to
social menu, will now call it in header.
Template- also adds in the ability for social, calls the div and
assigns ids